Resolve rebasing artifacts, move RepeatableKeyboard to Infrastructure.Input
